include_server: Avoid evaluating None macro definitions
It's possible that the macro definition is None (representing an empty
definition, such as from an include guard). In these cases, we do not
want to call _EvalMacro on them, as it violates the assertion that the
definition should be a string or a tuple.

A new test case enforces this, with a change to the Makefile to turn
python asserts on.

This should be safe in all cases because null definitions should not be
used in an #include (which is of course what we ultimately care about).

Signed-off-by: Alex Karle <akarle@mathworks.com>