include_server: Increase the limt of incoming connections.
We are able to push up to 800 concurrent compiles with pump mode,
but to do so it's necessary for the include_server to buffer requests
instead of dropping them.

As of Linux 5.4 (released Nov 2019), the default maximum on Linux is
4096 (configured by `sysctl net.core.somaxconn`). Prior to 2.4.25, it
was hardcoded as a constant `SOMAXCONN` to 128 (see also listen(2) NOTES).

This patch aligns the include_server default with Linux's maximum. It
should be safe on older versions and macOS which may have a lower
maximum, as the behavior of listen(2) is to silently cap at the minimum
of the requested backlog and the maximum setting.

Signed-off-by: Alex Karle <akarle@mathworks.com>