Remove strsignal from signal handlers to avoid crash
As discussed in #563, we're seeing the distccd daemon routinely crashing
in CI due to calling signal unsafe functions in signal handlers.

I've reproduced the crash locally, seeing it crash on both
`strsignal(3)` as well as `rs_log_info` (due to its use of `sprintf(3)`).

It appears there was already a calculated choice to run the risk of
tracing in the handler, per the comment in cleanup.c:

     * If from_signal_handler (a boolean) is non-zero, it means that
     * we're being called from a signal handler, so we need to be
     * careful not to call malloc() or free() or any other functions
     * that might not be async-signal-safe.
     * (We do call the tracing functions, which is perhaps unsafe
     * because they call sprintf() if DISCC_VERBOSE=1 is enabled.
     * But in that case it is probably worth the very small risk
     * of a crash to get the full tracing output.)

It's unclear if we still want to take this risk (I've never witnessed
a problem on Linux) or what the breakdown of macOS crashes due to
strsignal vs sprintf might be, but the easiest way to move the needle
for now without losing logging entirely is to disable strsignal on all
platforms in signal handlers.

Signed-off-by: Alex Karle <akarle@mathworks.com>